+++ cmd/apply.go
@@ -33,12 +33,8 @@ func runApply(cmd *cobra.Command, args []string) error {
 	if flagProject != "" {
 		stack.Project = flagProject
 	}
-	runtimeName := stack.Runtime
-	if flagRuntime != "" {
-		runtimeName = flagRuntime
-	}
 
-	rt, err := newRuntime(runtimeName)
+	rt, err := runtimeFrom(stack)
 	if err != nil {
 		return err
 	}