+++ cmd/down.go
@@ -31,12 +31,8 @@ func runDown(cmd *cobra.Command, args []string) error {
 	if flagProject != "" {
 		stack.Project = flagProject
 	}
-	runtimeName := stack.Runtime
-	if flagRuntime != "" {
-		runtimeName = flagRuntime
-	}
 
-	runtime, err := newRuntime(runtimeName)
+	runtime, err := runtimeFrom(stack)
 	if err != nil {
 		return err
 	}